pyproject.toml
[project]
name = "db-encrypt"
version = "0.1.0"
description = "Encrypt or decrypt database columns over JDBC using AES-256-GCM"
readme = "README.md"
authors = [
    { name = "test", email = "test@test.com" }
]
requires-python = ">=3.10"
dependencies = [
    "cryptography>=46.0.6",
    "jaydebeapi>=1.2.3",
    "jpype1>=1.6.0",
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
]

[project.scripts]
db-encrypt = "db_encrypt.cli:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
]
